Fix conversion error - add static cast (#98)
When compiling with -Werror using clang 18, there is an error: implicit
conversion changes signedness: 'int' to 'uint32_t' (aka 'unsigned int')
[-Werror,-Wsign-conversion]

+++ mctp_wrapper.hpp
@@ -39,7 +39,7 @@ struct DeviceID
     {
     }
     constexpr DeviceID(LocalEID eidVal, NetworkID nwid) :
-        id((nwid << 8) | eidVal)
+        id(static_cast<uint32_t>((nwid << 8) | eidVal))
     {
     }
     uint32_t id;
